podripper.sh: Leave only error logs
Removing status logs is more like UNIX-style when "no news is good
news". I haven't seen any errors in the script for a long time, so these
logs are less useful now. And trying the RCMP recording three days a
week when only one is likely to be successful would generate a lot of
unnecessary logs.