Use the newest of the matching upstream podcasts
If there are multiple podcasts that are "published" within one day
upstream, the script now picks the newest of them. This happened with
the RCMP podcast this week when episode 764 was officially published
last week, but it caused an incorrect RSS feed, so after publishing
episode 765, the previous one has the `pubDate` from this week as well
for some reason.