Remove the possibly partial re-encoded rip if ffmpeg fails
This is to prevent the strange situation that happened today with the
rcmp recording: `ffmpeg` got really slow and was still processing the
file hours later; this was probably due to bad networking because the
file is saved directly on S3 now. I terminated the process and thus both
the partially re-encoded and the source file ended up in the final
location and `rssgen` used only the partial file. With this change, the
possibly partial file from `ffmpeg` is removed if present.