Get rid of one-use variables when calculating `dateString`
This left-to-right flow is closer to the original one, although it's
necessarily more complicated due to the types involved. (It's
insteresting that every one of the three steps uses a different operator
to join it: `&`, `<$>` and `>>=`). This refactoring is arguably less
readable, but it eliminates the possibility of using an incorrect
variable.