Fix the changing time in rss item's title due to DST
This was spotted by the fact that my gPodder's extension failed to
remove the last March 2022 Radio-T episode because the expected filename
had changed (based on the episode's title). The episode's downloaded
filename is `sr_program_2022_03_26_21_50_18_enc.mp3`, and it was
recorded at 21:50 EET (UTC+2), so the original title was "2022-03-26
21:50:18 +0200 / Радио-Т 799". However after DST happened, its title
changed to "2022-03-26 20:50:18 +0200 / Радио-Т 799": note the incorrect
hour 20 instead of 21.

The fix is to get the timezone at that time (this already worked) and
then convert the local time to UTC using that timezone instead of the
current one because the source of the error was an incorrect UTC time:

```
λ> localTimeToUTC currentTZ localTime
2022-03-26 18:50:18 UTC
```