Use small reconnect delay ever since a successful rip
Testing this (more likely, even implementing this) required some
refactoring, namely the replacement of the infinite loop using
`repeatForever` with a more explicitly terminated loop using
`shouldRepeat`, because I couldn't add a `StateT Any m` inside the loop
and make the types work. The not necessarily infinite loop feels like a
workaround for testing, but maybe it's for the better this way.