Use the configured closest item time interval
The "default" value is left as 24 hours as before, but RCMP's interval
is only 10 hours because there is a regular recording on Friday night
at around 22:00 EET and there is an occasional recording on Saturday at
around 13:30 EET, so the default 24 hours assigns the Friday's title to
the Saturday's recording as well, which is incorrect. The 10 hour
interval should avoid this while still assigning the correct title for
Fridays.