Copy `rssgen`'s `src` and `test` files into `ripper`'s `RSSGen` dirs
All the tests pass.

The new `RSSGen` namespace is to have some separation of the two sets of
sources and tests (at least, initially). The `package.yaml` is updated
to include all the necessary dependencies. The new modules have caused a
bunch of warnings, so those warnings are temporarily disabled and must
be re-enabled and fixed shortly.

After starting this work, I thought of putting the modules into a
separate library (an "internal library" in [cabal][]) and test target,
however I'm not sure of big benefits of that now. The single program
will use all these modules.

[cabal]: https://cabal.readthedocs.io/en/stable/cabal-package.html#sublibs