Don't show `git` commit in program version
I think this is what causes the project to be built in the `pre-commit`
hook (that is, the `stack` commands aren't very fast) even if there are
no changes in the Haskell project. This happens due to `.git` changes:

```
$ stack build --fast
ripper-0.8.0: unregistering (local file changes: .git/HEAD .git/index .git/packed-refs .git/refs/heads/podripper_rewrite)
ripper> build (lib + exe)
…
```

Maybe, this isn't enough and the dependency on `githash` does this.