Fix not recording of non-ATP streams
The logic in the script was incorrect: it always used the ATP's status
before starting a recording, which is absolutely wrong for non-ATP
streams. Thus, the RCMP recording never started.

This change is to keep the livestream checks for ATP, and for other
streams, set the `STREAM_IS_LIVE` flag directly.