Ensure there is a sleep before a retry
It took me a while to get to the buildable state. It does work, but I
had to change `ST` to `IO` because there is no `MonadTime (ST s)`
instance (and I don't know how to create it, unlike `MonadTime IO`, for
which there is `MonadIO`, and there is no `STT` for `ST`).

The result somehow feels messy and ugly, and I'm not sure whether this
is the right approach…