Add logging to `runUntil`
It's made the function less pretty, but it will be good to see the
debugging information when running the program.

Why is `now` logged manually? It seems that `runStderrLoggingT` doesn't
add timestamps by default (or at all), and time is important in this
function. `systemd` does add timestamps on its own, but there is no
`systemd` when testing the program locally.

An SRP alternative would be to add a wrapper around `MonadTime` to log
its calls, but that could only log start and end of sleeping and
fetching the current time, which doesn't cover all logic in `runUntil`.