Return the entire response from `getFile`
As I understand, this change is necessary to be able to replace the
downloader function in the cached HTTP implementation (to be
implemented). Otherwise, it can't check the status code and headers.

This made `PollHTTPSpec` slightly more complicated because it needs to
`import` the internal module to create a `Response`.