Remove `PollHTTPSpec`
It's pretty much useless in the current version. Also, the
implementation of `pollHTTP` will be to add `runUntil`, so testing it
would be a more integration test repeating the cases in `RunUntilSpec`,
although more complex and specific to HTTP. In addition to that, I have
failed to implement a "decorator" in the MTL style, which would add
caching to `MonadDownload` and would be a `MonadDownload` itself`, so
I'm going to remove most/all of the custom monads since there will be
only one implementation of each.