Fix applying cache information to requests
Turns out an `ETag` was used with the `If-Modified-Since` header and a
`LastModified` was used with the `If-None-Match` header, which is
completely the opposite from how it should be. How did I come to this?!

Worse, this was discovered when debugging a not-updating ATP feed, only
after adding some logs, which showed a `400 Bad Request`.