Accept only one RSS filename in rssgen
There is not a big reason to support many at once, at least not in the
`podripper` itself. On the other hand, knowing the one filename enables
us to change the `shake` directory correspondingly in order to avoid
database locking issues when multiple podripper processes start (and
run) at the same time.