Remove support for `_src` rips
They were a result of a failed `ffmpeg` command. Now that the step has
been changed to the local processing, there is no such fallback. Yes,
there can be an exception processing the rips… but the exceptions aren't
handled at the moment (we'll see if there are any in production).