Use a vector to store MP3 frames
Appending to a list caused the program memory usage to go up while
ripping and used a lot of GC (~3× from the program's MUT time); it must
be because the list was copied for each append, which happened very
often.  Prepending a list made the CPU usage much better, but memory was
still high.

This first change isn't enough and in fact makes things work (similar to
the original list usage). I need to try using a mutable vector (it has a
more complicated write interface) and/or an unboxed vector.