Update the stack resolver to work with HLS
HLS for GHC 9.4.7 isn't provided anymore (I'm not sure what caused it,
could be a new version of ghcup 0.1.40.0?!), so HLS wouldn't work in
vim. `ghcup list -t ghc` shows which GHC versions are "hls-powered"; the
latest for now is 9.8.2, but there is no stack LTS for that version, so
I picked the next one, 9.6.4 for LTS 22.20 ("published 9 months ago").
The update and required fixes were pretty painless this time.